move nested class to separate files
